requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
sqlalchemy>=2.0
asyncpg>=0.29
alembic>=1.13
fastapi-users>=13.0
fastapi-users-db-sqlalchemy>=6.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
email-validator>=2.1
python-dotenv>=1.0

